requirements.txt
mutagen
pyyaml